+++ src/main/java/org/noostak/auth/dto/google/GoogleAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package org.noostak.auth.dto.kakao;
+package org.noostak.auth.dto.google;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
