+++ src/main/java/org/noostak/auth/dto/google/GoogleTokenRequest.java
@@ -1,4 +1,4 @@
-package org.noostak.auth.dto.kakao;
+package org.noostak.auth.dto.google;
 
 import lombok.Getter;
